requirements.txt
pandas
numpy
requests
beautifulsoup4
python-dotenv
yfinance
transformers
torch
scikit-learn
tqdm
